feat: refactored trainer to improve project modularity
