+++ predictor/pyproject.toml
@@ -19,6 +19,7 @@ dev = [
     "bandit>=1.8.0",
     "mypy>=1.13.0",
     "pytest>=8.3.5",
+    "pytest-cov>=6.2.1",
     "pytest-mock>=3.14.0",
     "ruff>=0.12.1",
 ]
