+++ predictor/tests/test_utils.py
@@ -1,6 +1,5 @@
 """Tests for the utility functions."""
 
-import os
 from typing import Any
 from unittest.mock import MagicMock, patch
