+++ trainer/pyproject.toml
@@ -20,7 +20,13 @@ dependencies = [
 ]
 
 [dependency-groups]
-dev = ["bandit>=1.8.0", "mypy>=1.13.0", "ruff>=0.8.2", "pytest>=8.0.0"]
+dev = [
+    "bandit>=1.8.0",
+    "mypy>=1.13.0",
+    "ruff>=0.8.2",
+    "pytest>=8.0.0",
+    "pytest-cov>=6.2.1",
+]
 
 [tool.ruff]
 line-length = 80
